EOS-20956-NWportHealthView: moved network_interface to /platforms/server dir, minor changes.
