+++ low-level/framework/platforms/server/network_interface.py
@@ -17,7 +17,7 @@ import os
 
 from framework.utils.conf_utils import (
     Conf, SSPL_CONF, SYSFS_PATH, SYSTEM_INFORMATION)
-from framework.platforms.node.error import InterfaceError
+from framework.platforms.server.error import InterfaceError
 
 
 class NetworkInterface:
